[test] refactor: consolidate the context-parallel tests into two files
Five files for 20 cases was too granular, and the split did not follow a
meaningful axis: the accelerator-config and parallel-state files were near
mirror images testing the same three constraints at two layers, and each
cross-referenced the other in its docstring to explain itself.

Regroup along the axis that matters. test_dsv4_cp_gates.py covers the three
admission gates a run passes in order -- TrainingArguments, ParallelState, and
the model allow-list -- so their relationship is stated once instead of by
cross-reference. test_dsv4_cp_sharding.py covers the mechanism: window
ownership arithmetic plus the collectives that move those shards, which now
share one canonical fixture. The per-rank compressed row counts the collectives
gather were previously hand-written as a literal in both files; they are now
defined once, next to the arithmetic that produces them.

No cases added or removed: the same 20 still collect and pass.
